Adding error condition tests for conditional steps
